+++ src/generator/angular-file-type.model.ts
@@ -2,6 +2,9 @@ export const NG_FILE_TYPE = {
   module: 'module',
   service: 'service',
   component: 'component',
+  directive: 'directive',
+  pipe: 'pipe',
+  route: 'route',
 } as const;
 
 export const NG_FILE_TYPES: NgFileType[] = Object.keys(